+++ efgs/uploader/src/main/java/pl/gov/mc/protegosafe/efgs/EfgsDiagnosisKeysUploader.java
@@ -1,13 +1,13 @@
 package pl.gov.mc.protegosafe.efgs;
 
-import eu.interop.federationgateway.model.EfgsProto;
 import lombok.AccessLevel;
 import lombok.experimental.FieldDefaults;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang3.RandomStringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import pl.gov.mc.protegosafe.efgs.repository.DiagnosisKeysRepository;
+import pl.gov.mc.protegosafe.efgs.repository.FailedDiagnosisKeysRepository;
+import pl.gov.mc.protegosafe.efgs.repository.model.DiagnosisKey;
 import pl.gov.mc.protegosafe.efgs.secret.CloudBackendConfig;
 import pl.gov.mc.protegosafe.efgs.utils.BatchSignatureUtils;
 
@@ -15,20 +15,13 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
-
 @Slf4j
 @Service
 @FieldDefaults(makeFinal = true, level = AccessLevel.PRIVATE)
-public class EfgsDiagnosisKeysUploader {
+public class EfgsDiagnosisKeysUploader extends DiagnosisKeysUploader implements DiagnosisKeysProcessor {
 
-    private static final int MAX_DIAGNOSIS_KEY_BATCH = 1000;
-
-    String nbbsCert;
-    EfgsFakeDiagnosisKeysFactory efgsFakeDiagnosisKeysFactory;
-    EfgsProtoDiagnosisKeyBatchFactory efgsProtoDiagnosisKeyBatchFactory;
-    BatchSignatureUtils batchSignatureUtils;
-    HttpUploader httpUploader;
-    DiagnosisKeysRepository diagnosisKeysRepository;
+    private final DiagnosisKeysRepository diagnosisKeysRepository;
+    private final FailedDiagnosisKeysRepository failedDiagnosisKeysRepository;
 
     @Autowired
     EfgsDiagnosisKeysUploader(
@@ -37,69 +30,46 @@ public class EfgsDiagnosisKeysUploader {
             BatchSignatureUtils batchSignatureUtils,
             HttpUploader httpUploader,
             DiagnosisKeysRepository diagnosisKeysRepository,
+            FailedDiagnosisKeysRepository failedDiagnosisKeysRepository,
             CloudBackendConfig cloudBackendConfig) {
-        this.nbbsCert = cloudBackendConfig.getNbbsCert();
-        this.efgsFakeDiagnosisKeysFactory = efgsFakeDiagnosisKeysFactory;
-        this.efgsProtoDiagnosisKeyBatchFactory = efgsProtoDiagnosisKeyBatchFactory;
-        this.batchSignatureUtils = batchSignatureUtils;
-        this.httpUploader = httpUploader;
+        super(efgsFakeDiagnosisKeysFactory, efgsProtoDiagnosisKeyBatchFactory, batchSignatureUtils, httpUploader, cloudBackendConfig);
         this.diagnosisKeysRepository = diagnosisKeysRepository;
+        this.failedDiagnosisKeysRepository = failedDiagnosisKeysRepository;
     }
 
-    void process() {
-
+    public void process() {
         log.info("started uploading keys to efgs");
 
-        Map<String, DiagnosisKey> idsWithDiagnosisKeys = getDocuments();
-        List<DiagnosisKey> diagnosisKeyBatch = createDiagnosisKeyList(idsWithDiagnosisKeys);
+        final Map<String, DiagnosisKey> idsWithDiagnosisKeys = diagnosisKeysRepository.getLimitedDiagnosisKeys();
+        final List<DiagnosisKey> diagnosisKeyBatch = idsWithDiagnosisKeys
+                .values()
+                .stream()
+                .collect(Collectors.toUnmodifiableList());
 
         if (diagnosisKeyBatch.isEmpty()) {
             log.info("No data to upload");
             return;
         }
 
-        List<DiagnosisKey> filledCollection = efgsFakeDiagnosisKeysFactory.fillFakesDiagnosisKeys(diagnosisKeyBatch, MAX_DIAGNOSIS_KEY_BATCH);
-
-        log.info("Processing upload, original keys: {}, fake keys: {}", diagnosisKeyBatch.size(), filledCollection.size() - diagnosisKeyBatch.size());
-
-        EfgsProto.DiagnosisKeyBatch batch = efgsProtoDiagnosisKeyBatchFactory.create(filledCollection);
-
-        byte[] bytes = batchSignatureUtils.generateBytesToVerify(batch);
-        SignatureGenerator signatureGenerator = new SignatureGenerator(nbbsCert);
-        String signatureForBytes = signatureGenerator.getSignatureForBytes(bytes);
-
-        String randomBatchTag = getRandomBatchTag();
-
-        boolean isSuccessResponse = httpUploader.uploadDiagnosisKeyBatch(batch, randomBatchTag, signatureForBytes);
+        final boolean isSuccessResponse = signAndUpload(diagnosisKeyBatch, false);
 
         if (!isSuccessResponse) {
-            processFailedUploading(diagnosisKeyBatch);
-
+            failedDiagnosisKeysRepository.saveFailedUploadingDiagnosisKeys(diagnosisKeyBatch);
+        } else {
+            log.info("Uploaded successfully");
         }
-        idsWithDiagnosisKeys.keySet()
+
+        idsWithDiagnosisKeys
+                .keySet()
                 .forEach(this::removeDocuments);
     }
 
+    @Override
+    public Boolean isApplicable(Mode mode) {
+        return Mode.UPLOAD_KEYS.equals(mode);
+    }
 
     private void removeDocuments(String documentId) {
         diagnosisKeysRepository.removeDocument(documentId);
-    }
-
-    private void processFailedUploading(List<DiagnosisKey> diagnosisKeys) {
-        diagnosisKeysRepository.saveFailedUploadingDiagnosisKeys(diagnosisKeys);
-    }
-
-    private String getRandomBatchTag() {
-        return RandomStringUtils.random(10, true, true);
-    }
-
-    private Map<String ,DiagnosisKey> getDocuments() {
-        return diagnosisKeysRepository.getLimitedDiagnosisKeys();
-    }
-
-    private List<DiagnosisKey> createDiagnosisKeyList(Map<String ,DiagnosisKey> documents) {
-        return documents.values()
-                .stream()
-                .collect(Collectors.toUnmodifiableList());
     }
 }