+++ efgs/uploader/src/main/java/pl/gov/mc/protegosafe/efgs/Mode.java
@@ -1,0 +1,5 @@
+package pl.gov.mc.protegosafe.efgs;
+
+public enum Mode {
+    UPLOAD_KEYS, RETRY_KEYS
+}