+++ efgs/uploader/src/main/java/pl/gov/mc/protegosafe/efgs/Properties.java
@@ -15,4 +15,5 @@ public class Properties {
   String efgsApiUrl;
   String projectId;
   Long diagnosisKeysFetchDelayFromRepository;
+  Mode mode = Mode.UPLOAD_KEYS;
 }