+++ functions/src/functions/freeTest/generateSubscriptionCode/index.ts
@@ -3,6 +3,7 @@ import {validateApiTokenAndIp} from "../../ipAndApiTokenValidator";
 import {generateCode} from "../../code/codeGenerator";
 import returnBadRequestResponse from "../../returnBadRequestResponse";
 import moment = require("moment");
+import config from "../../../config";
 
 const {log} = require("firebase-functions/lib/logger");
 const DELETE_LIFETIME = 48 * 60 * 60;
@@ -16,7 +17,10 @@ const generateSubscriptionCode = async (request: functions.Request, response: fu
             return returnBadRequestResponse(response);
         }
 
-        const {code, id} = await generateCode(moment().unix() + DELETE_LIFETIME);
+        const expiryTime = moment().unix() + config.code.lifetime * 60;
+        const deleteTime = moment().unix() + DELETE_LIFETIME;
+
+        const {code, id} = await generateCode(expiryTime, deleteTime);
 
         log(`generated lab test code, return code: 201`);