+++ functions/src/functions/jobs/backupTranslations/index.ts
@@ -1,20 +1,20 @@
 import axios from "axios";
-import config from "../../../config";
+import {secretManager} from "../../../services";
 import {saveFileInStorage} from "../storageSaver";
 import resolveRepository from "./Repository";
 
-const listLanguagesCodes = async (): Promise<Array<string>> => {
+const listLanguagesCodes = async (repository: any): Promise<Array<string>> => {
 
-    const response = await resolveRepository().post('', 'action=list_languages');
+    const response = await repository.post('', 'action=list_languages');
 
     const {list} = response.data;
 
     return list.map((item: any) => item.code)
 };
 
-const generateUrl = async (language: string): Promise<string> => {
+const generateUrl = async (language: string, repository: any): Promise<string> => {
 
-    const response = await resolveRepository().post('', `action=export&type=key_value_json&language=${language}`);
+    const response = await repository.post('', `action=export&type=key_value_json&language=${language}`);
 
     const {item} = response.data;
 
@@ -25,22 +25,25 @@ const throwError = (exception: string) => {
     throw new Error(exception)
 };
 
-const saveFile = (json: any, language: string) => {
+const saveFile = async (json: any, language: string) => {
     const fileName = `${new Date().getTime()}_${language}.json`;
     const compress = true;
-    saveFileInStorage(json.data, fileName, config.buckets.archive, compress)
+
+    const {archive} = await secretManager.getConfig('buckets');
+
+    saveFileInStorage(json.data, fileName, archive, compress)
         .catch(throwError)
 };
 
 const fetchFile = (url: string, language: string) => {
     axios.get(url)
         .catch(throwError)
-        .then(json => saveFile(json, language))
+        .then(json => async () => await saveFile(json, language))
         .catch(throwError)
 };
 
-const processLanguage = (language: string) => {
-    generateUrl(language)
+const processLanguage = (language: string, repository: any) => {
+    generateUrl(language, repository)
         .catch(throwError)
         .then(url => fetchFile(url, language))
         .catch(throwError)
@@ -49,8 +52,10 @@ const processLanguage = (language: string) => {
 const backupTranslations = async () => {
 
     try {
-        const languagesCodes = await listLanguagesCodes();
-        languagesCodes.forEach(processLanguage);
+        const {token, projectId} = await secretManager.getConfig('backupTranslations');
+        const repository = resolveRepository(token, projectId);
+        const languagesCodes = await listLanguagesCodes(repository);
+        languagesCodes.forEach(language => processLanguage(language, repository));
     } catch (exception) {
         throw new Error(exception);
     }